% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_braces}
\alias{expand_braces}
\title{Bash-style brace expansion}
\usage{
expand_braces(string)
}
\arguments{
\item{string}{input character vector}
}
\value{
A character vector
}
\description{
\code{expand_braces} provides partial support for
Bash-style brace expansion.
}
\examples{
  expand_braces("Foo{A..F}")
  expand_braces("Foo{01..10}")
  expand_braces("Foo{A..F}{1..5..2}")
  expand_braces("Foo{a..f..2}{-01..5}")
}
